HV-2025 Remove OSGi testing and support
