+++ documentation/src/test/java/org/hibernate/validator/referenceguide/chapter09/MultiClassLoaderScriptEvaluatorFactory.java
@@ -4,7 +4,7 @@
  * License: Apache License, Version 2.0
  * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
  */
-package org.hibernate.validator.osgi.scripting;
+package org.hibernate.validator.referenceguide.chapter09;
 
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
