--- osgi/felixtest/src/test/java/com/example/cdi/MyBean.java
@@ -1,19 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package com.example.cdi;
-
-import jakarta.enterprise.context.RequestScoped;
-
-@RequestScoped
-public class MyBean {
-
-	public void doDefault(@ValidNumber String number) {
-	}
-
-	public void doAll9(@ValidNumber("all9") String number) {
-	}
-}
