--- osgi/integrationtest/src/test/java/com/example/Customer.java
@@ -1,20 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package com.example;
-
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.Min;
-
-/**
- * @author Gunnar Morling
- */
-public class Customer {
-
-	@Min(1)
-	@DecimalMin("1.00")
-	private final int status = 0;
-}
