--- osgi/integrationtest/src/test/java/com/example/Order.java
@@ -1,16 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package com.example;
-
-/**
- * @author Gunnar Morling
- */
-public class Order {
-
-	@SuppressWarnings("unused")
-	private final String name = null;
-}
