--- osgi/integrationtest/src/test/java/com/example/constraint/ValidNameValidator.java
@@ -1,26 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package com.example.constraint;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-/**
- * @author Gunnar Morling
- *
- */
-public class ValidNameValidator implements ConstraintValidator<ValidName, String> {
-
-	@Override
-	public void initialize(ValidName constraintAnnotation) {
-	}
-
-	@Override
-	public boolean isValid(String value, ConstraintValidatorContext context) {
-		return false;
-	}
-}
