--- osgi/integrationtest/src/test/java/com/example/constraintvalidator/MustMatch.java
@@ -1,31 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package com.example.constraintvalidator;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Documented
-@Constraint(validatedBy = { })
-@Target({ METHOD, FIELD })
-@Retention(RUNTIME)
-public @interface MustMatch {
-	String message() default "{com.example.constraintvalidator.MustMatch.message}";
-
-	Class<?>[] groups() default { };
-
-	Class<? extends Payload>[] payload() default { };
-
-	String value();
-}
