--- osgi/integrationtest/src/test/java/com/example/constraintvalidator/MustMatchValidator.java
@@ -1,30 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package com.example.constraintvalidator;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-
-public class MustMatchValidator implements ConstraintValidator<MustMatch, String> {
-
-	private String match;
-
-	@Override
-	public void initialize(MustMatch constraintAnnotation) {
-		this.match = constraintAnnotation.value();
-	}
-
-	@Override
-	public boolean isValid(String value, ConstraintValidatorContext context) {
-		if ( value == null ) {
-			return true;
-		}
-
-		return value.equals( match );
-	}
-}
