--- osgi/integrationtest/src/test/resources/META-INF/services/jakarta.validation.ConstraintValidator
@@ -1,1 +1,0 @@
-com.example.constraintvalidator.MustMatchValidator
