--- osgi/integrationtest/src/test/resources/META-INF/validation.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Hibernate Validator, declare and validate application constraints
-  ~
-  ~ License: Apache License, Version 2.0
-  ~ See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
-  -->
-<validation-config
-        xmlns="https://jakarta.ee/xml/ns/validation/configuration"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="https://jakarta.ee/xml/ns/validation/configuration https://jakarta.ee/xml/ns/validation/validation-configuration-3.0.xsd"
-        version="3.0">
-
-    <constraint-validator-factory>
-        com.example.ExampleConstraintValidatorFactory
-    </constraint-validator-factory>
-
-    <constraint-mapping>META-INF/constraint-mappings.xml</constraint-mapping>
-</validation-config>
