--- osgi/integrationtest/src/test/resources/ValidationMessages.properties
@@ -1,7 +1,0 @@
-#
-# Hibernate Validator, declare and validate application constraints
-#
-# License: Apache License, Version 2.0
-# See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
-#
-com.example.RetailOrder.name.message=Not a valid retail order name
