--- osgi/karaf-features/src/main/features/features.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Hibernate Validator, declare and validate application constraints
-  ~
-  ~ License: Apache License, Version 2.0
-  ~ See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
-  -->
-<features xmlns="http://karaf.apache.org/xmlns/features/v1.4.0"
-    name="hibernate-validator-osgi-features"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://karaf.apache.org/xmlns/features/v1.4.0">
-
-    <feature name="hibernate-validator" version="${project.version}">
-        <bundle>mvn:org.hibernate.validator/hibernate-validator/${project.version}</bundle>
-        <bundle>mvn:jakarta.validation/jakarta.validation-api/${version.jakarta.validation-api}</bundle>
-
-        <bundle>mvn:org.jboss.logging/jboss-logging/${version.org.jboss.logging.jboss-logging}</bundle>
-        <bundle>mvn:com.fasterxml/classmate/${version.com.fasterxml.classmate}</bundle>
-        <bundle>mvn:jakarta.el/jakarta.el-api/${version.jakarta.el-api}</bundle>
-        <bundle>mvn:org.glassfish.expressly/expressly/${version.org.glassfish.expressly}</bundle>
-    </feature>
-    <feature name="hibernate-validator-joda-time" version="${project.version}">
-        <feature>hibernate-validator</feature>
-        <bundle>mvn:joda-time/joda-time/${version.joda-time}</bundle>
-    </feature>
-    <feature name="hibernate-validator-javax-money" version="${project.version}">
-        <feature>hibernate-validator</feature>
-        <bundle>mvn:javax.money/money-api/${version.javax.money}</bundle>
-        <bundle>mvn:org.javamoney/moneta/${version.org.javamoney.moneta}</bundle>
-    </feature>
-    <feature name="hibernate-validator-groovy" version="${project.version}">
-        <feature>hibernate-validator</feature>
-        <bundle>mvn:org.codehaus.groovy/groovy-all/${version.org.codehaus.groovy}</bundle>
-    </feature>
-    <feature name="hibernate-validator-paranamer" version="${project.version}">
-        <feature prerequisite="true">wrap</feature>
-        <feature>hibernate-validator</feature>
-        <bundle>wrap:mvn:com.thoughtworks.paranamer/paranamer/${version.com.thoughtworks.paranamer}</bundle>
-    </feature>
-</features>
