--- osgi/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Hibernate Validator, declare and validate application constraints
-  ~
-  ~ License: Apache License, Version 2.0
-  ~ See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.hibernate.validator</groupId>
-        <artifactId>hibernate-validator-parent</artifactId>
-        <version>8.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>hibernate-validator-osgi</artifactId>
-    <packaging>pom</packaging>
-
-    <name>Hibernate Validator OSGi Modules</name>
-    <description>Hibernate Validator OSGi modules aggregator</description>
-
-    <properties>
-        <hibernate-validator-parent.path>..</hibernate-validator-parent.path>
-    </properties>
-
-    <modules>
-        <module>karaf-features</module>
-        <module>integrationtest</module>
-    </modules>
-    
-    <profiles>
-        <profile>
-            <id>testWithJdk8</id>
-            <activation>
-                <property>
-                    <name>java-version.test.release</name>
-                    <value>8</value>
-                </property>
-            </activation>
-            <modules>
-                <!--
-                Only enable this test on JDK 8 as it is failing on JDK 9+.
-                See https://hibernate.atlassian.net/browse/HV-1478 for more information about the failure.
-                -->
-                <module>felixtest</module>
-            </modules>
-        </profile>
-    </profiles>
-</project>
