+++ pom.xml
@@ -211,7 +211,6 @@
         <version.org.apache.karaf>4.2.0</version.org.apache.karaf>
         <version.org.ops4j.pax.exam>4.12.0</version.org.ops4j.pax.exam>
         <version.org.ops4j.pax.url>2.5.4</version.org.ops4j.pax.url>
-        <version.org.osgi.core>6.0.0</version.org.osgi.core>
         <version.fish.payara>5.2020.2</version.fish.payara>
         <version.fish.payara.arquillian>2.3.1</version.fish.payara.arquillian>
         <!-- Used in the Karaf tests -->
@@ -461,11 +460,6 @@
                 <groupId>net.bytebuddy</groupId>
                 <artifactId>byte-buddy</artifactId>
                 <version>${version.net.bytebuddy.byte-buddy}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.osgi</groupId>
-                <artifactId>org.osgi.core</artifactId>
-                <version>${version.org.osgi.core}</version>
             </dependency>
             <dependency>
                 <groupId>org.apache.logging.log4j</groupId>
